+++ source/TuiStringUtils.h
@@ -54,7 +54,7 @@ inline std::string string_format(const std::string fmt, ...) {
 
 inline std::wstring convertUtf8ToWide(const std::string& str)
 {
-    int count = MultiByteToWideChar(CP_UTF8, 0, str.c_str(), str.length(), NULL, 0);
+    int count = MultiByteToWideChar(CP_UTF8, 0, str.c_str(), (int)str.length(), NULL, 0);
     std::wstring wstr(count, 0);
     MultiByteToWideChar(CP_UTF8, 0, str.c_str(), str.length(), &wstr[0], count);
     return wstr;
